feat(extractor): Separate content in content and package.
