+++ handler/pipelines/extractor/content.py
@@ -20,787 +20,14 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 Should there be a need for contact the electronic mail
 `handler <at> gabrielfontenelle.com` can be used.
 """
-from datetime import datetime
-from io import IOBase
-from tarfile import TarFile, TarError
-from zipfile import BadZipFile, ZipFile
-
-from ..hasher import CRC32Hasher
 from tinytag import TinyTag
 
-from py7zr import SevenZipFile
-from py7zr.exceptions import Bad7zFile
-from rarfile import BadRarFile, RarFile, NotRarFile
-
 from .extractor import Extractor
-from .. import Pipeline
-
-from ...exception import ValidationError
 
 __all__ = [
     'AudioMetadataFromContentExtractor',
-    'ContentExtractor',
     'MimeTypeFromContentExtractor',
-    'SevenZipCompressedFilesFromContentExtractor',
-    'RarCompressedFilesFromContentExtractor',
-    'TarCompressedFilesFromContentExtractor',
-    'ZipCompressedFilesFromContentExtractor',
 ]
-
-
-class ContentExtractor(Extractor):
-    """
-    Extractor class with focus to processing information from file's content.
-    This class was created to allow parsing and extraction of data designated to
-    internal files.
-    """
-
-    extensions = None
-    """
-    Attribute to store allowed extensions for use in `validator`.
-    This attribute should be override in children classes.
-    """
-    compressor_class = None
-    """
-    Attribute to store the current class of compressor for use in `content_buffer` and `decompress` methods.
-    This attribute should be override in children classes.
-    """
-    stopper = True
-    """
-    Variable that define if this class used as processor should stop the pipeline.
-    """
-
-    class ContentBuffer(IOBase):
-        """
-        Class to allow consumption of buffer in a lazy way.
-        This class should be override in children of ContentExtractor to
-        implementation of method read().
-        """
-
-        source_file_object = None
-        """
-        Attribute to store the related file object that has the buffer for the compressed content.
-        """
-        compressor = None
-        """
-        Attribute to store the class of the compressor able to uncompress the content.  
-        """
-        filename = None
-        """
-        Attribute to store the name of file that should be extract for this content.
-        """
-
-        def __init__(self, source_file_object, compressor_class, internal_file_filename):
-            """
-            Method to initiate the object saving the data required to allow decompressing and reading content
-            for specific file.
-            """
-            self.source_file_object = source_file_object
-            self.compressor = compressor_class
-            self.filename = internal_file_filename
-
-        def read(self, *args, **kwargs) -> str:
-            """
-            Method to read the content of the object initiating the buffer if not exists.
-            This method should be overwritten in child class.
-            """
-            raise NotImplementedError(f"Method read of ContentBuffer should be override in child class "
-                                      f"{self.__class__.__name__}.")
-
-        def seek(self, *args, **kwargs):
-            """
-            Method to seek the content in the buffer.
-            Buffer must exist for this method to work, else no action will be taken.
-            """
-            if not hasattr(self, "buffer"):
-                return
-
-            return self.buffer.seek(*args, **kwargs)
-
-        def seekable(self):
-            """
-            Method to verify if buffer is seekable.
-            Buffer must exist for this method to work, else no action will be taken.
-            """
-
-            if not hasattr(self, "buffer"):
-                return False
-
-            return self.buffer.seekable()
-
-        def close(self):
-            """
-            Method to close the buffer.
-            Buffer must exist for this method to work, else no action will be taken.
-            """
-            if not hasattr(self, "buffer"):
-                return
-
-            self.buffer.close()
-            delattr(self, "buffer")
-
-    @classmethod
-    def validate(cls, file_object):
-        """
-        Method to validate if content can be extract to given extension.
-        """
-        if cls.extensions is None:
-            raise NotImplementedError(f"The attribute extensions is not overwritten in child class {cls.__name__}")
-
-        # The ValidationError should be captured in children classes else it will not register as an error and
-        # the pipeline will break.
-        if file_object.extension not in cls.extensions:
-            raise ValidationError(f"Extension {file_object.extension} not allowed in validate for class {cls.__name__}")
-
-    @classmethod
-    def decompress(cls, file_object, overrider: bool, **kwargs: dict):
-        """
-        Method to uncompress the content from a file_object.
-        This method must be override in child class.
-        """
-        raise NotImplementedError("Method extract_content must be overwritten on child class.")
-
-    @classmethod
-    def content_buffer(cls, file_object, internal_file_name):
-        """
-        Method to create a buffer pointing to the uncompressed content.
-        This method must work lazily, extracting the content only when the buffer is read.
-        This method must be override in child class.
-        """
-        raise NotImplementedError("Method content_buffer must be overwritten on child class.")
-
-
-class MastrokaFilesFromContentExtractor(ContentExtractor):
-    """
-    Class to extract internal files from mka, mkv files.
-    """
-
-
-class PSDLayersFromContentExtractor(ContentExtractor):
-    """
-    Class to extract internal files from PSD files.
-    """
-
-
-class TarCompressedFilesFromContentExtractor(ContentExtractor):
-    """
-    Class to extract internal files from tar.gz or tar.bz files.
-    As gzip and bzip compression don`t provide a manifest, as it is not an archive format just a compression algorithm,
-    the tarfile library should be used to try to extract information of files with compression .bz and .gz as its the
-    most common usage of tar with those compressors.
-    """
-
-    extensions = ['gz', 'tar', 'bz']
-    """
-    Attribute to store allowed extensions for use in `validator`.
-    """
-    compressor = TarFile
-    """
-    Attribute to store the current class of compressor for use in `content_buffer` and `decompress` methods.
-    """
-
-    @classmethod
-    def content_buffer(cls, file_object, internal_file_name):
-        """
-        Method to create a buffer pointing to the uncompressed content.
-        This method must work lazily, extracting the content only when the buffer is read.
-        """
-
-        class TarContentBuffer(cls.ContentBuffer):
-            """
-            Class to allow consumption of buffer in a lazy way.
-            """
-
-            def read(self, *args, **kwargs) -> str:
-                """
-                Method to read the content of the object initiating the buffer if not exists.
-                """
-                if not hasattr(self, "buffer"):
-                    # Instantiate the buffer of inner content
-                    compressed = self.compressor(fileobj=self.source_file_object.buffer)
-
-                    self.buffer = compressed.extractfile(member=self.filename)
-
-                return self.buffer.read(*args, **kwargs)
-
-        return TarContentBuffer(file_object, cls.compressor_class, internal_file_name)
-
-    @classmethod
-    def decompress(cls, file_object, overrider: bool, **kwargs: dict):
-        """
-        Method to uncompress the content from a file_object.
-        """
-        try:
-            cls.validate(file_object)
-
-            # We don't need to create the directory because the extractor will create it if not exists.
-            extraction_path = kwargs.pop("decompress_to")
-
-            # We don't need to reset the buffer before calling it, because it will be reset
-            # if already cached. The next time property buffer is called it will reset again.
-            with cls.compressor(fileobj=file_object.buffer) as compressed_file:
-
-                # targets as None will extract all data, overwriting existing ones.
-                targets = None
-
-                if not overrider:
-                    targets = []
-                    for filename in compressed_file.getnames():
-                        # Avoid files beginning with `..` or `/` for security reason.
-                        if filename[0] == '/' or filename[0:2] == '..':
-                            continue
-
-                        if not file_object.storage.exists(
-                                file_object.storage.join(extraction_path, filename)
-                        ):
-                            targets.append(filename)
-
-                    # Avoid calling the extractor if the list is empty.
-                    if not targets:
-                        return True
-
-                # Concurrent extract file in external file system using a custom pathlib.Path
-                # with accessor informed by storage of file_object.
-                compressed_file.extractall(path=file_object.storage.get_pathlib_path(extraction_path), members=targets)
-
-        except (BadZipFile, ValidationError):
-            return False
-
-        return True
-
-    @classmethod
-    def extract(cls, file_object, overrider: bool, **kwargs: dict):
-        """
-        Method to extract the information necessary from a file_object.
-        """
-        try:
-            cls.validate(file_object)
-
-            file_system = file_object.storage
-            file_class = file_object.__class__
-
-            # We don't need to reset the buffer before calling it, because it will be reset
-            # if already cached. The next time property buffer is called it will reset again.
-            with cls.compressor(fileobj=file_object.buffer) as compressed_object:
-
-                for internal_file in compressed_object.getmembers():
-                    # Skip directories
-                    if internal_file.isdir():
-                        continue
-
-                    # Skip duplicate only if not choosing to override.
-                    if internal_file.name in file_object._content_files and not overrider:
-                        continue
-
-                    # Create file object for internal file
-                    internal_file_object = file_class(
-                        path=internal_file.name,
-                        extract_data_pipeline=Pipeline(
-                            'handler.pipelines.extractor.FilenameAndExtensionFromPathExtractor',
-                            'handler.pipelines.extractor.MimeTypeFromFilenameExtractor',
-                        ),
-                        file_system_handler=file_system
-                    )
-
-                    # Update creation and modified date
-                    internal_file_object.create_date = internal_file.mtime
-                    internal_file_object.update_date = internal_file.mtime
-
-                    # Update size of file
-                    internal_file_object.size = internal_file.size
-
-                    # Update hash generating the hash file and adding its content
-                    if internal_file.chksum:
-                        hash_file = CRC32Hasher.create_hash_file(
-                            object_to_process=file_object,
-                            digested_hex_value=internal_file.chksum
-                        )
-                        internal_file_object.hashes['crc32'] = internal_file.chksum, hash_file, CRC32Hasher
-
-                    # Set up action to be extracted instead of to save.
-                    internal_file_object._actions.to_extract()
-
-                    # Set up content pointer to internal file using content_buffer
-                    internal_file_object.content = cls.content_buffer(file_object, internal_file.name)
-
-                    # Set up metadata for internal file
-                    internal_file_object.meta.hashable = False
-                    internal_file_object.meta.internal = True
-
-                    # Add internal file as File object to file.
-                    file_object._content_files[internal_file.name] = internal_file_object
-
-            # Update metadata and actions.
-            file_object.meta.packed = True
-            file_object._actions.listed()
-
-            return True
-
-        except (TarError, ValidationError):
-            return False
-
-
-class ZipCompressedFilesFromContentExtractor(ContentExtractor):
-    """
-    Class to extract internal files from rar files.
-    """
-
-    extensions = ['zip']
-    """
-    Attribute to store allowed extensions for use in `validator`.
-    """
-    compressor = ZipFile
-    """
-    Attribute to store the current class of compressor for use in `content_buffer` and `decompress` methods.
-    """
-
-    @classmethod
-    def content_buffer(cls, file_object, internal_file_name):
-        """
-        Method to create a buffer pointing to the uncompressed content.
-        This method must work lazily, extracting the content only when the buffer is read.
-        """
-
-        class ZipContentBuffer(cls.ContentBuffer):
-            """
-            Class to allow consumption of buffer in a lazy way.
-            """
-
-            def read(self, *args, **kwargs) -> str:
-                """
-                Method to read the content of the object initiating the buffer if not exists.
-                """
-                if not hasattr(self, "buffer"):
-                    # Instantiate the buffer of inner content
-                    compressed = self.compressor(file=self.source_file_object.buffer)
-
-                    self.buffer = compressed.open(name=self.filename)
-
-                return self.buffer.read(*args, **kwargs)
-
-        return ZipContentBuffer(file_object, cls.compressor_class, internal_file_name)
-
-    @classmethod
-    def decompress(cls, file_object, overrider: bool, **kwargs: dict):
-        """
-        Method to uncompress the content from a file_object.
-        """
-        try:
-            cls.validate(file_object)
-
-            # We don't need to create the directory because the extractor will create it if not exists.
-            extraction_path = kwargs.pop("decompress_to")
-
-            # We don't need to reset the buffer before calling it, because it will be reset
-            # if already cached. The next time property buffer is called it will reset again.
-            with cls.compressor(file=file_object.buffer) as compressed_file:
-
-                # targets as None will extract all data, overwriting existing ones.
-                targets = None
-
-                if not overrider:
-                    targets = []
-                    for filename in compressed_file.namelist():
-                        # Avoid files beginning with `..` or `/` for security reason.
-                        if filename[0] == '/' or filename[0:2] == '..':
-                            continue
-
-                        if not file_object.storage.exists(
-                            file_object.storage.join(extraction_path, filename)
-                        ):
-                            targets.append(filename)
-
-                    # Avoid calling the extractor if the list is empty.
-                    if not targets:
-                        return True
-
-                # Concurrent extract file in external file system using a custom pathlib.Path
-                # with accessor informed by storage of file_object.
-                compressed_file.extractall(path=file_object.storage.get_pathlib_path(extraction_path), members=targets)
-
-        except (BadZipFile, ValidationError):
-            return False
-
-        return True
-
-    @classmethod
-    def extract(cls, file_object, overrider: bool, **kwargs: dict):
-        """
-        Method to extract the information necessary from a file_object.
-        """
-        try:
-            cls.validate(file_object)
-
-            file_system = file_object.storage
-            file_class = file_object.__class__
-
-            # We don't need to reset the buffer before calling it, because it will be reset
-            # if already cached. The next time property buffer is called it will reset again.
-            with cls.compressor(file=file_object.buffer) as compressed_object:
-
-                for internal_file in compressed_object.infolist():
-                    # Skip directories and symbolic link
-                    if internal_file.is_dir():
-                        continue
-
-                    # Skip duplicate only if not choosing to override.
-                    if internal_file.filename in file_object._content_files and not overrider:
-                        continue
-
-                    # Create file object for internal file
-                    internal_file_object = file_class(
-                        path=internal_file.filename,
-                        extract_data_pipeline=Pipeline(
-                            'handler.pipelines.extractor.FilenameAndExtensionFromPathExtractor',
-                            'handler.pipelines.extractor.MimeTypeFromFilenameExtractor',
-                        ),
-                        file_system_handler=file_system
-                    )
-
-                    # Update creation and modified date. Zip don't store the created date, only the modified one.
-                    # To avoid problem the created date will be consider the same as modified.
-                    internal_file_object.create_date = datetime(*internal_file.date_time)
-                    internal_file_object.update_date = internal_file_object.create_date
-
-                    # Update size of file
-                    internal_file_object.size = internal_file.file_size
-
-                    # Update hash generating the hash file and adding its content
-                    hash_file = CRC32Hasher.create_hash_file(
-                        object_to_process=file_object,
-                        digested_hex_value=internal_file.CRC
-                    )
-                    internal_file_object.hashes['crc32'] = internal_file.CRC, hash_file, CRC32Hasher
-
-                    # Set up action to be extracted instead of to save.
-                    internal_file_object._actions.to_extract()
-
-                    # Set up content pointer to internal file using content_buffer
-                    internal_file_object.content = cls.content_buffer(file_object, internal_file.filename)
-
-                    # Set up metadata for internal file
-                    internal_file_object.meta.hashable = False
-                    internal_file_object.meta.internal = True
-
-                    # Add internal file as File object to file.
-                    file_object._content_files[internal_file.filename] = internal_file_object
-
-            # Update metadata and actions.
-            file_object.meta.packed = True
-            file_object._actions.listed()
-
-            return True
-
-        except (BadZipFile, ValidationError):
-            return False
-
-
-class RarCompressedFilesFromContentExtractor(ContentExtractor):
-    """
-    Class to extract internal files from rar files.
-    """
-
-    extensions = ['rar']
-    """
-    Attribute to store allowed extensions for use in `validator`.
-    """
-    compressor = RarFile
-    """
-    Attribute to store the current class of compressor for use in `content_buffer` and `decompress` methods.
-    """
-
-    @classmethod
-    def content_buffer(cls, file_object, internal_file_name):
-        """
-        Method to create a buffer pointing to the uncompressed content.
-        This method must work lazily, extracting the content only when the buffer is read.
-        """
-
-        class RarContentBuffer(cls.ContentBuffer):
-            """
-            Class to allow consumption of buffer in a lazy way.
-            """
-
-            def read(self, *args, **kwargs) -> str:
-                """
-                Method to read the content of the object initiating the buffer if not exists.
-                """
-                if not hasattr(self, "buffer"):
-                    # Instantiate the buffer of inner content
-                    compressed = self.compressor(file=self.source_file_object.buffer)
-
-                    self.buffer = compressed.open(name=self.filename)
-
-                return self.buffer.read(*args, **kwargs)
-
-        return RarContentBuffer(file_object, cls.compressor_class, internal_file_name)
-
-    @classmethod
-    def decompress(cls, file_object, overrider: bool, **kwargs: dict):
-        """
-        Method to uncompress the content from a file_object.
-        """
-        try:
-            cls.validate(file_object)
-
-            # We don't need to create the directory because the extractor will create it if not exists.
-            extraction_path = kwargs.pop("decompress_to")
-
-            # We don't need to reset the buffer before calling it, because it will be reset
-            # if already cached. The next time property buffer is called it will reset again.
-            with cls.compressor(file=file_object.buffer) as compressed_file:
-
-                # targets as None will extract all data, overwriting existing ones.
-                targets = None
-
-                if not overrider:
-                    targets = []
-                    for filename in compressed_file.namelist():
-                        # Avoid files beginning with `..` or `/` for security reason.
-                        if filename[0] == '/' or filename[0:2] == '..':
-                            continue
-
-                        if not file_object.storage.exists(
-                            file_object.storage.join(extraction_path, filename)
-                        ):
-                            targets.append(filename)
-
-                    # Avoid calling the extractor if the list is empty.
-                    if not targets:
-                        return True
-
-                # Concurrent extract file in external file system using a custom pathlib.Path
-                # with accessor informed by storage of file_object.
-                compressed_file.extractall(path=file_object.storage.get_pathlib_path(extraction_path), members=targets)
-
-        except (BadRarFile, NotRarFile, ValidationError):
-            return False
-
-        return True
-
-    @classmethod
-    def extract(cls, file_object, overrider: bool, **kwargs: dict):
-        """
-        Method to extract the information necessary from a file_object.
-        """
-        try:
-            cls.validate(file_object)
-
-            file_system = file_object.storage
-            file_class = file_object.__class__
-
-            # We don't need to reset the buffer before calling it, because it will be reset
-            # if already cached. The next time property buffer is called it will reset again.
-            with cls.compressor(file=file_object.buffer) as compressed_object:
-
-                for internal_file in compressed_object.infolist():
-                    # Skip directories and symbolic link
-                    if internal_file.is_dir() or internal_file.is_symlink():
-                        continue
-
-                    # Skip duplicate only if not choosing to override.
-                    if internal_file.filename in file_object._content_files and not overrider:
-                        continue
-
-                    # Create file object for internal file
-                    internal_file_object = file_class(
-                        path=internal_file.filename,
-                        extract_data_pipeline=Pipeline(
-                            'handler.pipelines.extractor.FilenameAndExtensionFromPathExtractor',
-                            'handler.pipelines.extractor.MimeTypeFromFilenameExtractor',
-                        ),
-                        file_system_handler=file_system
-                    )
-
-                    # Update creation and modified date
-                    internal_file_object.create_date = internal_file.ctime
-                    internal_file_object.update_date = internal_file.mtime
-
-                    # Update size of file
-                    internal_file_object.size = internal_file.file_size
-
-                    # Update hash generating the hash file and adding its content
-                    if internal_file.CRC:
-                        hash_file = CRC32Hasher.create_hash_file(
-                            object_to_process=file_object,
-                            digested_hex_value=internal_file.CRC
-                        )
-                        internal_file_object.hashes['crc32'] = internal_file.CRC, hash_file, CRC32Hasher
-
-                    # Set up action to be extracted instead of to save.
-                    internal_file_object._actions.to_extract()
-
-                    # Set up content pointer to internal file using content_buffer
-                    internal_file_object.content = cls.content_buffer(file_object, internal_file.filename)
-
-                    # Set up metadata for internal file
-                    internal_file_object.meta.hashable = False
-                    internal_file_object.meta.internal = True
-
-                    # Add internal file as File object to file.
-                    file_object._content_files[internal_file.filename] = internal_file_object
-
-            # Update metadata and actions.
-            file_object.meta.packed = True
-            file_object._actions.listed()
-
-            return True
-
-        except (BadRarFile, NotRarFile, ValidationError):
-            return False
-
-
-class SevenZipCompressedFilesFromContentExtractor(ContentExtractor):
-    """
-    Class to extract internal files from 7z files.
-    """
-
-    extensions = ['7z']
-    """
-    Attribute to store allowed extensions for use in `validator`.
-    """
-    compressor = SevenZipFile
-    """
-    Attribute to store the current class of compressor for use in `content_buffer` and `decompress` methods.
-    """
-
-    @classmethod
-    def content_buffer(cls, file_object, internal_file_name):
-        """
-        Method to create a buffer pointing to the uncompressed content.
-        This method must work lazily, extracting the content only when the buffer is read.
-        """
-        class SevenZipContentBuffer(cls.ContentBuffer):
-            """
-            Class to allow consumption of buffer in a lazy way.
-            """
-
-            def read(self, *args, **kwargs) -> str:
-                """
-                Method to read the content of the object initiating the buffer if not exists.
-                """
-                if not hasattr(self, "buffer"):
-                    # Instantiate the buffer of inner content
-                    compressed = self.compressor(file=self.source_file_object.buffer)
-
-                    self.buffer = next(iter(compressed.read(targets=[self.filename]).values()))
-
-                return self.buffer.read(*args, **kwargs)
-
-        return SevenZipContentBuffer(file_object, cls.compressor_class, internal_file_name)
-
-    @classmethod
-    def decompress(cls, file_object, overrider: bool, **kwargs: dict):
-        """
-        Method to uncompress the content from a file_object.
-        """
-        try:
-            cls.validate(file_object)
-
-            # We don't need to create the directory because the extractor will create it if not exists.
-            extraction_path = kwargs.pop("decompress_to")
-
-            # We don't need to reset the buffer before calling it, because it will be reset
-            # if already cached. The next time property buffer is called it will reset again.
-            with cls.compressor(file=file_object.buffer) as compressed_file:
-
-                # targets as None will extract all data, overwriting existing ones.
-                targets = None
-
-                if not overrider:
-                    targets = []
-                    for filename in compressed_file.getnames():
-                        # Avoid files beginning with `..` or `/` for security reason.
-                        if filename[0] == '/' or filename[0:2] == '..':
-                            continue
-
-                        if not file_object.storage.exists(
-                            file_object.storage.join(extraction_path, filename)
-                        ):
-                            targets.append(filename)
-
-                    # Avoid calling the extractor if the list is empty.
-                    if not targets:
-                        return True
-
-                # Concurrent extract file in external file system using a custom pathlib.Path
-                # with accessor informed by storage of file_object.
-                compressed_file.extract(path=file_object.storage.get_pathlib_path(extraction_path), targets=targets)
-
-        except (ValidationError, Bad7zFile):
-            return False
-
-        return True
-
-    @classmethod
-    def extract(cls, file_object, overrider: bool, **kwargs: dict):
-        """
-        Method to extract the information necessary from a file_object.
-        """
-        try:
-            cls.validate(file_object)
-
-            file_system = file_object.storage
-            file_class = file_object.__class__
-
-            # We don't need to reset the buffer before calling it, because it will be reset
-            # if already cached. The next time property buffer is called it will reset again.
-            with cls.compressor(file=file_object.buffer) as compressed_object:
-
-                for internal_file in compressed_object.list():
-                    # Skip directories
-                    if internal_file.is_directory:
-                        continue
-
-                    # Skip duplicate only if not choosing to override.
-                    if internal_file.filename in file_object._content_files and not overrider:
-                        continue
-
-                    # Create file object for internal file
-                    internal_file_object = file_class(
-                        path=internal_file.filename,
-                        extract_data_pipeline=Pipeline(
-                            'handler.pipelines.extractor.FilenameAndExtensionFromPathExtractor',
-                            'handler.pipelines.extractor.MimeTypeFromFilenameExtractor',
-                        ),
-                        file_system_handler=file_system
-                    )
-
-                    # Update creation and modified date
-                    internal_file_object.create_date = internal_file.creationtime
-                    internal_file_object.update_date = internal_file.creationtime
-
-                    # Update size of file
-                    internal_file_object.size = internal_file.uncompressed
-
-                    # Update hash generating the hash file and adding its content
-                    hash_file = CRC32Hasher.create_hash_file(
-                        object_to_process=file_object,
-                        digested_hex_value=internal_file.crc32
-                    )
-                    internal_file_object.hashes['crc32'] = internal_file.crc32, hash_file, CRC32Hasher
-
-                    # Set up action to be extracted instead of to save.
-                    internal_file_object._actions.to_extract()
-
-                    # Set up content pointer to internal file using content_buffer
-                    internal_file_object.content = cls.content_buffer(file_object, internal_file.filename)
-
-                    # Set up metadata for internal file
-                    internal_file_object.meta.hashable = False
-                    internal_file_object.meta.internal = True
-
-                    # Add internal file as File object to file.
-                    file_object._content_files[internal_file.filename] = internal_file_object
-
-            # Update metadata and actions.
-            file_object.meta.packed = True
-            file_object._actions.listed()
-
-            return True
-
-        except (Bad7zFile, ValidationError):
-            return False
 
 
 class VideoMetadataFromContentExtractor(Extractor):
@@ -822,6 +49,10 @@ class ImageMetadataFromContentExtractor(Extractor):
 
 
 class AudioMetadataFromContentExtractor(Extractor):
+    """
+    # what its
+    # usage of class
+    """
 
     @classmethod
     def extract(cls, file_object, overrider: bool, **kwargs: dict):
